requirements.txt
BeautifulSoup==3.2.0
Django==1.2.4
Numeric==24.2
South==0.7.3
distribute==0.6.14
django-cbv==0.1.5
django-haystack==1.1.0
mercurial==1.7.2
palb==0.1.1
psycopg2==2.4-beta2
pycurl==7.19.0
redis==2.2.0
unittest2==0.5.1
virtualenv==1.5.1
wsgiref==0.1.2
xapian-haystack==1.1.5beta
markdown
typogrify
sorl-thumbnail==11.01
PIL
django-sentry
